Adds nagging for destinationDir property in AbstractCompile task
- Getters and setters now warn about removal in Gradle 9, not 8.
- Tests updated to use replacement property wherever possible.
- BackwardCompatibilityOutputDirectoryCOnvention moved out of AbstractCompile task into JavaBasePlugin, where it is applied to all compile tasks.
