+++ subprojects/language-java/src/integTest/groovy/org/gradle/api/tasks/compile/JavaCompileTaskIntegrationTest.groovy
@@ -20,7 +20,7 @@ import org.gradle.integtests.fixtures.AbstractIntegrationSpec
 import spock.lang.Issue
 
 class JavaCompileTaskIntegrationTest extends AbstractIntegrationSpec {
-    def "task does nothing when no configuration applied"() {
+    def "task does nothing when only minimal configuration applied"() {
         buildFile << """
             // No plugins applied
             task compile(type: JavaCompile)
