+++ subprojects/language-java/src/test/groovy/org/gradle/api/tasks/compile/JavaCompileTest.groovy
@@ -127,7 +127,7 @@ class JavaCompileTest extends AbstractProjectBuilderSpec {
 
     def "spec is configured using the toolchain compiler in-process using the current jvm as toolchain and sets release"() {
         def javaCompile = project.tasks.create("compileJava", JavaCompile)
-        javaCompile.setDestinationDir(new File("tmp"))
+        javaCompile.destinationDirectory = new File("tmp")
         def javaHome = Jvm.current().javaHome
         def metadata = Mock(JavaInstallationMetadata)
         def compiler = Mock(JavaCompiler)
@@ -151,7 +151,7 @@ class JavaCompileTest extends AbstractProjectBuilderSpec {
     @Issue('https://bugs.openjdk.java.net/browse/JDK-8139607')
     def "spec is configured using the toolchain compiler in-process using the current jvm as toolchain and does not set release for Java 9"() {
         def javaCompile = project.tasks.create("compileJava", JavaCompile)
-        javaCompile.setDestinationDir(new File("tmp"))
+        javaCompile.destinationDirectory = new File("tmp")
         def javaHome = Jvm.current().javaHome
         def metadata = Mock(JavaInstallationMetadata)
         def compiler = Mock(JavaCompiler)
@@ -174,7 +174,7 @@ class JavaCompileTest extends AbstractProjectBuilderSpec {
 
     def "spec is configured using the toolchain compiler in-process using the current jvm as toolchain and set source and target compatibility"() {
         def javaCompile = project.tasks.create("compileJava", JavaCompile)
-        javaCompile.setDestinationDir(new File("tmp"))
+        javaCompile.destinationDirectory = new File("tmp")
         def javaHome = Jvm.current().javaHome
         def metadata = Mock(JavaInstallationMetadata)
         def compiler = Mock(JavaCompiler)
