+++ subprojects/language-jvm/src/main/java/org/gradle/api/tasks/compile/AbstractCompile.java
@@ -15,10 +15,8 @@
  */
 package org.gradle.api.tasks.compile;
 
-import org.gradle.api.file.Directory;
 import org.gradle.api.file.DirectoryProperty;
 import org.gradle.api.file.FileCollection;
-import org.gradle.api.internal.GeneratedSubclasses;
 import org.gradle.api.model.ReplacedBy;
 import org.gradle.api.provider.Provider;
 import org.gradle.api.tasks.Classpath;
@@ -29,8 +27,6 @@ import org.gradle.internal.deprecation.DeprecationLogger;
 import org.gradle.work.DisableCachingByDefault;
 
 import java.io.File;
-import java.lang.reflect.Method;
-import java.util.concurrent.Callable;
 
 /**
  * The base class for all JVM-based language compilation tasks.
@@ -44,7 +40,6 @@ public abstract class AbstractCompile extends SourceTask {
 
     public AbstractCompile() {
         this.destinationDirectory = getProject().getObjects().directoryProperty();
-        this.destinationDirectory.convention(getProject().getProviders().provider(new BackwardCompatibilityOutputDirectoryConvention()));
     }
 
     /**
@@ -82,17 +77,17 @@ public abstract class AbstractCompile extends SourceTask {
      *
      * @return The destination directory.
      *
-     * @deprecated Use {@link #getDestinationDirectory()} instead. This method will be removed in Gradle 8.0.
+     * @deprecated Use {@link #getDestinationDirectory()} instead. This method will be removed in Gradle 9.0.
      */
     @ReplacedBy("destinationDirectory")
     @Deprecated
     public File getDestinationDir() {
         // Used in Kotlin plugin - needs updating there and bumping the version first. Followup with https://github.com/gradle/gradle/issues/16783
-        /*DeprecationLogger.deprecateProperty(AbstractCompile.class, "destinationDir")
+        DeprecationLogger.deprecateProperty(AbstractCompile.class, "destinationDir")
             .replaceWith("destinationDirectory")
-            .willBeRemovedInGradle8()
+            .willBeRemovedInGradle9()
             .withUpgradeGuideSection(7, "compile_task_wiring")
-            .nagUser();*/
+            .nagUser();
 
         return destinationDirectory.getAsFile().getOrNull();
     }
@@ -102,13 +97,13 @@ public abstract class AbstractCompile extends SourceTask {
      *
      * @param destinationDir The destination directory. Must not be null.
      *
-     * @deprecated Use {@link #getDestinationDirectory()}.set() instead. This method will be removed in Gradle 8.0.
+     * @deprecated Use {@link #getDestinationDirectory()}.set() instead. This method will be removed in Gradle 9.0.
      */
     @Deprecated
     public void setDestinationDir(File destinationDir) {
         DeprecationLogger.deprecateProperty(AbstractCompile.class, "destinationDir")
             .replaceWith("destinationDirectory")
-            .willBeRemovedInGradle8()
+            .willBeRemovedInGradle9()
             .withUpgradeGuideSection(7, "compile_task_wiring")
             .nagUser();
 
@@ -121,16 +116,16 @@ public abstract class AbstractCompile extends SourceTask {
      * @param destinationDir The destination directory. Must not be null.
      * @since 4.0
      *
-     * @deprecated Use {@link #getDestinationDirectory()}.set() instead. This method will be removed in Gradle 8.0.
+     * @deprecated Use {@link #getDestinationDirectory()}.set() instead. This method will be removed in Gradle 9.0.
      */
     @Deprecated
     public void setDestinationDir(Provider<File> destinationDir) {
         // Used by Android plugin. Followup with https://github.com/gradle/gradle/issues/16782
-        /*DeprecationLogger.deprecateProperty(AbstractCompile.class, "destinationDir")
+        DeprecationLogger.deprecateProperty(AbstractCompile.class, "destinationDir")
             .replaceWith("destinationDirectory")
-            .willBeRemovedInGradle8()
+            .willBeRemovedInGradle9()
             .withUpgradeGuideSection(7, "compile_task_wiring")
-            .nagUser();*/
+            .nagUser();
 
         this.destinationDirectory.set(getProject().getLayout().dir(destinationDir));
     }
@@ -171,45 +166,5 @@ public abstract class AbstractCompile extends SourceTask {
      */
     public void setTargetCompatibility(String targetCompatibility) {
         this.targetCompatibility = targetCompatibility;
-    }
-
-    /**
-     * Convention to fall back to the 'destinationDir' output for backwards compatibility with plugins that extend AbstractCompile and override the deprecated methods.
-     *
-     * TODO - move this into the class decoration
-     */
-    private class BackwardCompatibilityOutputDirectoryConvention implements Callable<Directory> {
-        private boolean recursiveCall;
-
-        @Override
-        public Directory call() throws Exception {
-            Method getter = GeneratedSubclasses.unpackType(AbstractCompile.this).getMethod("getDestinationDir");
-            if (getter.getDeclaringClass() == AbstractCompile.class) {
-                // Subclass has not overridden the getter, so ignore
-                return null;
-            }
-
-            // Subclass has overridden the getter, so call it
-
-            if (recursiveCall) {
-                // Already querying AbstractCompile.getDestinationDirectory()
-                // In that case, this convention should not be used.
-                return null;
-            }
-            recursiveCall = true;
-            File legacyValue;
-            try {
-                // This will call a subclass implementation of getDestinationDir(), which possibly will not call the overridden getter
-                // In the Kotlin plugin, the subclass manages its own field which will be used here.
-                legacyValue = getDestinationDir();
-            } finally {
-                recursiveCall = false;
-            }
-            if (legacyValue == null) {
-                return null;
-            } else {
-                return getProject().getLayout().getProjectDirectory().dir(legacyValue.getAbsolutePath());
-            }
-        }
     }
 }
